pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "oanda-fx-ml"
version = "0.1.0"
description = "FX price streaming and feature pipeline on OANDA v20"
authors = [{name = "Your Name", email = "you@example.com"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "python-dotenv",
  "oandapyV20",
  "pandas",
  "numpy",
  "scikit-learn"
]

[project.urls]
Homepage = "https://developer.oanda.com/"
